Gourmet_backend_flask/requirements.txt
flask
flask_cors
python-dotenv

psycopg2-binary